# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pharmacy Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List/search products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResult"
  /products/expiry:
    get:
      operationId: getExpiringProducts
      tags: [products]
      summary: Get expired and expiring products
      parameters:
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Expiring products grouped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpiryReport"

  # Sales
...
components:
  schemas:
...
    MessageResult:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # Auth
...
    Product:
      type: object
      required: [id, name, genericName, category, quantity, reorderLevel, purchasePrice, sellingPrice, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        genericName:
          type: string
        category:
          type: string
        barcode:
          type: ["string", "null"]
        batchNumber:
          type: ["string", "null"]
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        purchasePrice:
          type: number
        sellingPrice:
          type: number
        quantity:
          type: integer
        reorderLevel:
          type: integer
        expiryDate:
          type: ["string", "null"]
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, genericName, category, purchasePrice, sellingPrice, quantity, reorderLevel]
      properties:
        name:
          type: string
        genericName:
          type: string
        category:
          type: string
        barcode:
          type: string
        batchNumber:
          type: string
        supplierId:
          type: integer
        purchasePrice:
          type: number
        sellingPrice:
          type: number
        quantity:
          type: integer
        reorderLevel:
          type: integer
        expiryDate:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        genericName:
          type: string
        category:
          type: string
        barcode:
          type: string
        batchNumber:
          type: string
        supplierId:
          type: integer
        purchasePrice:
          type: number
        sellingPrice:
          type: number
        quantity:
          type: integer
        reorderLevel:
          type: integer
        expiryDate:
          type: string
    ExpiryReport:
      type: object
      required: [expired, expiringSoon]
      properties:
        expired:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        expiringSoon:
          type: array
          items:
            $ref: "#/components/schemas/Product"

    # Sale
...